requirements.txt
streamlit
faiss-cpu
openai
python-dotenv
pypdf
numpy
